cleanup: Final cleanup, move each file into their respective directory
