+++ CMakeLists.txt
@@ -3,4 +3,4 @@ project(project_01 C)
 
 set(CMAKE_C_STANDARD 11)
 
-add_executable(project_01 main.c main.h register.c register.h user.c user.h file.c file.h login.c login.h board.c board.h card.c card.h user_db.c user_db.h ui.c ui.h types.h board_db.c board_db.h user_menu.c user_menu.h id_array.c id_array.h global_db.c global_db.h utils.c utils.h board_menu.c board_menu.h user_search.c user_search.h card_array.c card_array.h card_menu.c card_menu.h)
+add_executable(project_01 main.c main.h menu/register_menu.c menu/register_menu.h model/user.c model/user.h util/file.c util/file.h menu/login_menu.c menu/login_menu.h model/board.c model/board.h model/card.c model/card.h db/user_db.c db/user_db.h util/types.h db/board_db.c db/board_db.h menu/user_menu.c menu/user_menu.h container/id_array.c container/id_array.h db/global_db.c db/global_db.h util/utils.c util/utils.h menu/board_menu.c menu/board_menu.h container/user_search.c container/user_search.h container/card_array.c container/card_array.h menu/card_menu.c menu/card_menu.h)
