+++ menu/user_menu.h
@@ -5,11 +5,11 @@
 #ifndef PROJECT_01_USER_MENU_H
 #define PROJECT_01_USER_MENU_H
 
-#include "user.h"
-#include "main.h"
-#include "utils.h"
-#include "board_db.h"
 #include "board_menu.h"
+#include "../model/user.h"
+#include "../util/utils.h"
+#include "../db/board_db.h"
+#include "../main.h"
 
 void createBoard(USER *user);
 void listBoards(USER *user);
