--- ui.h
@@ -1,16 +1,0 @@
-//
-// Created by User on 2021/04/13.
-//
-
-#ifndef PROJECT_01_UI_H
-#define PROJECT_01_UI_H
-
-#include <stdlib.h>
-#include <stdio.h>
-
-#include "user.h"
-
-void clearScreen();
-char* getUsername();
-
-#endif //PROJECT_01_UI_H
